+++ app/src/main/java/com/mapbox/maps/testapp/examples/AnimatedImageSourceActivity.kt
@@ -1,12 +1,12 @@
 package com.mapbox.maps.testapp.examples
 
-import android.content.Context
-import android.graphics.Bitmap
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
 import androidx.appcompat.app.AppCompatActivity
-import com.mapbox.maps.MapboxMap
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
+import com.mapbox.maps.Image
+import com.mapbox.maps.MapboxDelicateApi
 import com.mapbox.maps.Style
 import com.mapbox.maps.extension.style.layers.generated.rasterLayer
 import com.mapbox.maps.extension.style.sources.generated.ImageSource
@@ -17,6 +17,10 @@ import com.mapbox.maps.extension.style.style
 import com.mapbox.maps.testapp.R
 import com.mapbox.maps.testapp.databinding.ActivityAnimatedImagesourceBinding
 import com.mapbox.maps.testapp.utils.BitmapUtils.bitmapFromDrawableRes
+import com.mapbox.maps.toMapboxImage
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.launch
 
 /**
  * Load a raster image to a style using ImageSource and display it on a map as
@@ -24,15 +28,12 @@ import com.mapbox.maps.testapp.utils.BitmapUtils.bitmapFromDrawableRes
  */
 class AnimatedImageSourceActivity : AppCompatActivity() {
 
-  private val handler: Handler = Handler(Looper.getMainLooper())
-  private lateinit var mapboxMap: MapboxMap
-  private lateinit var runnable: Runnable
-
+  @OptIn(MapboxDelicateApi::class)
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     val binding = ActivityAnimatedImagesourceBinding.inflate(layoutInflater)
     setContentView(binding.root)
-    mapboxMap = binding.mapView.mapboxMap
+    val mapboxMap = binding.mapView.mapboxMap
     mapboxMap.loadStyle(
       style(style = Style.STANDARD) {
         +imageSource(ID_IMAGE_SOURCE) {
@@ -48,49 +49,27 @@ class AnimatedImageSourceActivity : AppCompatActivity() {
         +rasterLayer(ID_IMAGE_LAYER, ID_IMAGE_SOURCE) { }
       }
     )
-  }
-
-  override fun onStart() {
-    super.onStart()
+    val drawables: List<Image> = listOf(
+      bitmapFromDrawableRes(R.drawable.southeast_radar_0).toMapboxImage(),
+      bitmapFromDrawableRes(R.drawable.southeast_radar_1).toMapboxImage(),
+      bitmapFromDrawableRes(R.drawable.southeast_radar_2).toMapboxImage(),
+      bitmapFromDrawableRes(R.drawable.southeast_radar_3).toMapboxImage(),
+    )
+    var drawableIndex = 0
     mapboxMap.getStyle {
       val imageSource: ImageSource = it.getSourceAs(ID_IMAGE_SOURCE)!!
-      runnable = RefreshImageRunnable(applicationContext, imageSource, handler)
-      handler.postDelayed(runnable, 100)
-    }
-  }
-
-  override fun onStop() {
-    super.onStop()
-    if (::runnable.isInitialized) {
-      handler.removeCallbacks(runnable)
-    }
-  }
-
-  private class RefreshImageRunnable constructor(
-    appContext: Context,
-    private val imageSource: ImageSource,
-    private val handler: Handler
-  ) :
-    Runnable {
-    private val drawables: Array<Bitmap?> = arrayOfNulls(4)
-    private var drawableIndex: Int
-
-    override fun run() {
-      drawables[drawableIndex++]?.let { bitmap ->
-        imageSource.updateImage(bitmap)
-        if (drawableIndex > 3) {
-          drawableIndex = 0
+      // Create a new coroutine in the lifecycleScope
+      lifecycleScope.launch {
+        // repeatOnLifecycle launches the block in a new coroutine every time the
+        // lifecycle is in the STARTED state (or above) and cancels it when it's STOPPED.
+        repeatOnLifecycle(Lifecycle.State.STARTED) {
+          while (isActive) {
+            imageSource.updateImage(drawables[drawableIndex++])
+            drawableIndex %= drawables.size
+            delay(1000L)
+          }
         }
       }
-      handler.postDelayed(this, 1000)
-    }
-
-    init {
-      drawables[0] = bitmapFromDrawableRes(appContext, R.drawable.southeast_radar_0)
-      drawables[1] = bitmapFromDrawableRes(appContext, R.drawable.southeast_radar_1)
-      drawables[2] = bitmapFromDrawableRes(appContext, R.drawable.southeast_radar_2)
-      drawables[3] = bitmapFromDrawableRes(appContext, R.drawable.southeast_radar_3)
-      drawableIndex = 1
     }
   }