+++ sdk-base/src/main/java/com/mapbox/maps/ExtensionUtils.kt
@@ -7,7 +7,11 @@ import com.mapbox.bindgen.DataRef
 
 /**
  * Convert [Bitmap] to rendering engine [Image] instance.
+ *
+ * Important: This will allocate native memory outside the JVM heap. If possible, avoid using it
+ * inside loops/animations.
  */
+@MapboxDelicateApi
 fun Bitmap.toMapboxImage(): Image {
   if (config != Bitmap.Config.ARGB_8888) {
     throw IllegalArgumentException("Only ARGB_8888 bitmap config is supported!")